+++ packages/x-adapter-platform/src/platform.adapter.ts
@@ -1,6 +1,5 @@
 import { searchEndpointAdapter } from './endpoint-adapters/search.endpoint-adapter';
 import { PlatformAdapter } from './types/platform-adapter.types';
-import { empathizeEndpointAdapter } from './endpoint-adapters/empathize.endpoint-adapter';
 import { topClickedEndpointAdapter } from './endpoint-adapters/top-clicked.endpoint-adapter';
 import { nextQueriesEndpointAdapter } from './endpoint-adapters/next-queries.endpoint-adapter';
 import { relatedTagsEndpointAdapter } from './endpoint-adapters/related-tags.endpoint-adapter';
@@ -9,7 +8,6 @@ import { identifierResultsEndpointAdapter } from './endpoint-adapters/identifier
 import { taggingEndpointAdapter } from './endpoint-adapters/tagging.endpoint-adapter';
 
 export const platformAdapter: PlatformAdapter = {
-  empathize: empathizeEndpointAdapter,
   identifierResults: identifierResultsEndpointAdapter,
   nextQueries: nextQueriesEndpointAdapter,
   popularSearches: topClickedEndpointAdapter,
