+++ packages/x-adapter-platform/src/types/responses/identifier-results-response.model.ts
@@ -1,5 +1,10 @@
-import { PlatformResult } from '../models.types';
+import { PlatformResult } from './models/result.model';
 
+/**
+ * Response for the `identifier results` endpoint.
+ *
+ * @public
+ */
 export interface PlatformIdentifierResultsResponse {
   catalog: {
     content: PlatformResult[];
