+++ packages/x-adapter-platform/src/types/responses/index.ts
@@ -1,6 +1,8 @@
-export * from './empathize-response.model';
 export * from './identifier-results-response.model';
+export * from './models';
 export * from './next-queries-response.model';
+export * from './popular-searches.response.model';
+export * from './query-suggestions-response.model';
+export * from './recommendations-response.model';
 export * from './related-tags-response.model';
 export * from './search-response.model';
-export * from './top-clicked-response.model';
