+++ packages/x-adapter-platform/src/types/responses/models/banner.model.ts
@@ -1,0 +1,14 @@
+/**
+ * Banner model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformBanner {
+  id: string;
+  title: string;
+  url: string;
+  image_url: string;
+  tagging?: {
+    query: string;
+  };
+}
