+++ packages/x-adapter-platform/src/types/responses/models/index.ts
@@ -1,0 +1,9 @@
+export * from './banner.model';
+export * from './facet.model';
+export * from './next-query.model';
+export * from './promoted.model';
+export * from './redirection.model';
+export * from './related-tag.model';
+export * from './result.model';
+export * from './suggestion.model';
+export * from './tagging.model';
