+++ packages/x-adapter-platform/src/types/responses/models/next-query.model.ts
@@ -1,0 +1,10 @@
+/**
+ * Next query model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformNextQuery {
+  query: string;
+  source: 'ORGANIC' | 'CURATED';
+  position: number;
+}
