+++ packages/x-adapter-platform/src/types/responses/models/related-tag.model.ts
@@ -1,0 +1,11 @@
+/**
+ * Related tag model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformRelatedTag {
+  query: string;
+  tag: string;
+  source: 'ORGANIC' | 'CURATED';
+  position: number;
+}
