+++ packages/x-adapter-platform/src/types/responses/models/result.model.ts
@@ -1,0 +1,18 @@
+import { PlatformTagging } from './tagging.model';
+
+/**
+ * Result model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformResult {
+  name: string;
+  averageRating?: number;
+  id: string;
+  images?: string[];
+  image: string;
+  price: number;
+  sku?: string;
+  url: string;
+  tagging: PlatformTagging;
+}
