+++ packages/x-adapter-platform/src/types/responses/models/suggestion.model.ts
@@ -1,0 +1,8 @@
+/**
+ * Suggestion model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformSuggestion {
+  title_raw: string;
+}
