+++ packages/x-adapter-platform/src/types/responses/models/tagging.model.ts
@@ -1,0 +1,10 @@
+/**
+ * Tagging model for the `platform` API.
+ *
+ * @public
+ */
+export interface PlatformTagging {
+  add2cart: string;
+  checkout: string;
+  click: string;
+}
