+++ packages/x-adapter-platform/src/types/responses/next-queries-response.model.ts
@@ -1,5 +1,10 @@
-import { PlatformNextQuery } from '../models.types';
+import { PlatformNextQuery } from './models/next-query.model';
 
+/**
+ * Response for the `next queries` endpoint.
+ *
+ * @public
+ */
 export interface PlatformNextQueriesResponse {
   data: {
     nextqueries: PlatformNextQuery[];
