+++ packages/x-adapter-platform/src/types/responses/popular-searches.response.model.ts
@@ -1,0 +1,12 @@
+import { PlatformSuggestion } from './models/suggestion.model';
+
+/**
+ * Response for the `popular searches` endpoint.
+ *
+ * @public
+ */
+export interface PlatformPopularSearchesResponse {
+  topTrends: {
+    content: PlatformSuggestion[];
+  };
+}
