+++ packages/x-adapter-platform/src/types/responses/query-suggestions-response.model.ts
@@ -1,0 +1,13 @@
+import { PlatformSuggestion } from './models/suggestion.model';
+
+/**
+ * Response for the `query suggestions` endpoint.
+ *
+ * @public
+ */
+export interface PlatformQuerySuggestionsResponse {
+  topTrends: {
+    content: PlatformSuggestion[];
+    spellcheck?: string;
+  };
+}
