+++ packages/x-adapter-platform/src/types/responses/recommendations-response.model.ts
@@ -1,0 +1,13 @@
+import { PlatformResult } from './models/index';
+
+/**
+ * Response for the `recommendations` endpoint.
+ *
+ * @public
+ */
+export interface PlatformRecommendationsResponse {
+  topclicked: {
+    content: PlatformResult[];
+    numFound: number;
+  };
+}
