+++ packages/x-adapter-platform/src/types/responses/related-tags-response.model.ts
@@ -1,5 +1,10 @@
-import { PlatformRelatedTag } from '../models.types';
+import { PlatformRelatedTag } from './models/related-tag.model';
 
+/**
+ * Response for the `related tags` endpoint.
+ *
+ * @public
+ */
 export interface PlatformRelatedTagsResponse {
   data: {
     relatedtags: PlatformRelatedTag[];
