+++ packages/x-adapter-platform/src/types/responses/search-response.model.ts
@@ -1,14 +1,26 @@
-import {
-  Contentable,
-  PlatformBanner,
-  PlatformCatalog,
-  PlatformPromoted,
-  PlatformRedirection
-} from '../models.types';
+import { PlatformResult } from './models/result.model';
+import { PlatformBanner } from './models/banner.model';
+import { PlatformFacet } from './models/facet.model';
+import { PlatformPromoted } from './models/promoted.model';
+import { PlatformRedirection } from './models/redirection.model';
 
 export interface PlatformSearchResponse {
-  banner: Contentable<PlatformBanner>;
-  catalog: PlatformCatalog;
-  direct: Contentable<PlatformRedirection>;
-  promoted: Contentable<PlatformPromoted>;
+  banner: {
+    content: PlatformBanner[];
+  };
+  catalog: {
+    content: PlatformResult[];
+    facets: PlatformFacet[];
+    numFound: number;
+    spellchecked: string;
+    tagging: {
+      query: string;
+    };
+  };
+  direct: {
+    content: PlatformRedirection[];
+  };
+  promoted: {
+    content: PlatformPromoted[];
+  };
 }
